EDU-PATH - source code/EDU-PATH-BACKEND/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
flask = "^3.0.0"
Flask-SQLite = "^0.3.0"
Flask-SQLAlchemy = "^3.0.0"
Werkzeug = "^2.2.2"
Flask-RESTful = "^0.3.9"
matplotlib = "^3.5.3"
numpy = "^1.23.2"
Flask-Bcrypt = "^1.0.1"
Flask-Cors = "^3.0.10"
pytest = "^6.2.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"